requirements_complete.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0
xformers>=0.0.20
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
safetensors>=0.3.0
huggingface-hub>=0.16.0
scipy>=1.10.0
scikit-image>=0.20.0 